requirements.txt
django
Pillow
celery
redis
langchain
replicate
python-dotenv
cloudinary
langchain-google-genai